refactor: extract components repository and formatter
