+++ server/views/editor/components-response.php
@@ -2,19 +2,19 @@
 
 declare(strict_types=1);
 
+use Components\Formatter;
+use Components\Repository;
 use PDO;
-
-require_once __DIR__ . '/../../lib/components.php';
-require_once __DIR__ . '/../../lib/definitions.php';
 
 /**
  * @param array{message?: string|null, message_type?: string} $options
  */
 function components_render_fragments(\PDO $pdo, array $options = []): void
 {
-
-    $componentsTree = components_fetch_tree($pdo);
-    $componentsFlat = components_flatten_tree($componentsTree);
+    $formatter = new Formatter();
+    $repository = new Repository($pdo, $formatter);
+    $componentsTree = $repository->fetchTree();
+    $componentsFlat = $formatter->flattenTree($componentsTree);
     $definitionsTree = definitions_fetch_tree($pdo);
     $definitionsFlat = definitions_flatten_tree($definitionsTree);
